ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
openai==2.16.0
loguru==0.7.3
python-dotenv==1.2.1
